Updated script to use the new website at IAA-CSIC.
The script now parses the content of the new Drupal website that over
the last months has gradually replaced the old site. The advantage of
the new design is that all the information is publicly available, so
there is no need to access to the intranet -- which, by the way, now
requires an username and password, even if the connection is made from
a computer at the workplace.

This "General Alphabetical List" also includes associate researchers,
but not visiting ones. I've been confirmed that this could be easily
fixed, but it's not up to me -- somebody higher in the hierarchy will
have to ask for it to be done.