Added local cache — website is now accessed only once a week.
Due to the slow speed of the IAA website, the script now uses a local
cache of the alphabetical list (which is pickled to the filesystem),
so that the website is only accessed *once a week*. The -c options
allows the user to clear the local cache, forcing the connection to
the website and the updating of the cache.

I would like to thank Juan de Dios Santander for the suggestion.