Set default encoding to UTF-8
Use sys.setdefaultencoding() to switch from the default ASCII to
UTF-8. The use of this function is discouraged and, in fact, became a
no-op in Py3K, where the encoding is hard-wired to UTF-8 — but, until
we transition to Python 3, we can use this hack to prevent our script
from raising UnicodeEncodeError when its output is redirected. Thanks
to Juan de Dios Santander, who reported this problem.

Why we need sys.setdefaultencoding(“utf-8”) in a py script?
http://stackoverflow.com/q/3828723/184363