Allow more than one search term
It seems that some users pipe the output of the script to grep in
order to filter the results (e.g., `phone john | grep -i doe`). Add
this functionality to our code: if more than one search term is given
by the user, only those records that match all of them will be shown.